+++ R/vlaanderen_2015_colours.R
@@ -1,4 +1,4 @@
-#' A colour pallet according the Vlaanderen style guide version >= 2015.
+#' A colour pallet according the Flanders style guide version >= 2015.
 #'
 #' @export
 #' @param n The number of colours
@@ -9,8 +9,8 @@
 #' @importFrom assertthat assert_that is.count noNA
 vlaanderen_palette <- function(n) {
   palette <- c(
-    vl_black, vl_grey4, vl_lightgreen, vl_lightbrown, vl_lightblue, vl_lightred,
-    vl_darkgreen, vl_darkbrown, vl_darkblue, vl_darkred
+    vl_lightblue, vl_lightgreen, vl_darkblue, vl_darkbrown, vl_grey4,
+    vl_lightbrown, vl_lightred, vl_black, vl_darkred, vl_darkgreen
   )
   if (missing(n)) {
     n <- length(palette)